move storage implementations to custom-example
